Refactor: Extract save_geometry to io_writers.py
